added Installing ProBuilder link to index.md
User-generated bugs suggest some users cannot find the Installing ProBuilder link in the sidebar. I added another link to the installation steps to the About ProBuilder page to assist these users. A lot of packages include installation steps on the first page of the package's documentation, so this change will bring ProBuilder more in line with what user's expectations might be.