Tidying up TOC
I was using a couple of wrong files.

Also, I'm killing the Tools vs Actions page. First of all, we can't keep calling them actions because Actions is a category of Actions, and we can't do that to our poor users. Second, this page duplicates information that's better presented in its own context, such as creating meshes and the so-called actions. The conceptual difference between tool and action is really not something users need to worry themselves with.